feat: Make `record_v2` and `utils` tree-shakeable
