+++ js/package.json
@@ -45,7 +45,6 @@
     "@rollup/plugin-replace": "^5.0.5",
     "@rollup/plugin-terser": "^0.4.4",
     "@rollup/plugin-typescript": "^11.1.5",
-    "@solana/web3.js": "^1.87.6",
     "@tsconfig/recommended": "^1.0.3",
     "@types/bn.js": "^5.1.5",
     "@types/bs58": "^4.0.4",
@@ -58,6 +57,7 @@
     "jest": "^29.7.0",
     "mathjs": "^12.4.2",
     "prettier": "^3.1.0",
+    "rollup-plugin-multi-input": "^1.4.1",
     "rollup-plugin-visualizer": "^5.12.0",
     "ts-jest": "^29.1.1",
     "ts-node": "^10.9.1",
