+++ js/src/index.ts
@@ -1,7 +1,24 @@
 export * from "./bindings";
 export * from "./state";
 export * from "./twitter_bindings";
-export * from "./utils";
+
+export * from "./utils/check";
+export * from "./utils/deserializeReverse";
+export * from "./utils/findSubdomains";
+export * from "./utils/getAllDomains";
+export * from "./utils/getAllRegisteredDomains";
+export * from "./utils/getDomainKeySync";
+export * from "./utils/getDomainKeysWithReverses";
+export * from "./utils/getDomainPriceFromName";
+export * from "./utils/getHashedNameSync";
+export * from "./utils/getNameAccountKeySync";
+export * from "./utils/getPythFeedAccountKey";
+export * from "./utils/getReverseKeyFromDomainKey";
+export * from "./utils/getReverseKeySync";
+export * from "./utils/getTokenizedDomains";
+export * from "./utils/reverseLookup";
+export * from "./utils/reverseLookupBatch";
+
 export * from "./instructions";
 export * from "./nft";
 export { getDomainMint } from "./nft/name-tokenizer";
@@ -15,6 +32,14 @@ export * from "./resolve";
 export * from "./deprecated/utils";
 export * from "./error";
 export * from "./custom-bg";
-export * from "./record_v2";
+
+export * from "./record_v2/const";
+export * from "./record_v2/deserializeRecordV2Content";
+export * from "./record_v2/serializeRecordV2Content";
+export * from "./record_v2/getRecordV2";
+export * from "./record_v2/getRecordV2Key";
+export * from "./record_v2/getMultipleRecordsV2";
+export * from "./record_v2/verifyRightOfAssociation";
 export * from "./record_v2/utils";
+
 export * from "./devnet";
