+++ js/src/utils/check.ts
@@ -1,0 +1,7 @@
+import { SNSError } from "../error";
+
+export const check = <T extends SNSError>(bool: boolean, error: T) => {
+  if (!bool) {
+    throw error;
+  }
+};
