+++ js/tests/get-domain-price-from-name.test.ts
@@ -1,30 +1,30 @@
 require("dotenv").config();
 import { describe, test, expect } from "@jest/globals";
-import { getDomainPriceFromName } from "../src/utils";
+import { getDomainPriceFromName } from "../src/utils/getDomainPriceFromName";
 
 describe("getDomainPriceFromName", () => {
   test.each([
-    ['1', 750],
-    ['✅', 750],
-    ['요', 750],
-    ['👩‍👩‍👧', 750],
+    ["1", 750],
+    ["✅", 750],
+    ["요", 750],
+    ["👩‍👩‍👧", 750],
 
-    ['10', 700],
-    ['1✅', 700],
-    ['👩‍👩‍👧✅', 700],
-    ['독도', 700],
+    ["10", 700],
+    ["1✅", 700],
+    ["👩‍👩‍👧✅", 700],
+    ["독도", 700],
 
-    ['100', 640],
-    ['10✅', 640],
-    ['1독도', 640],
+    ["100", 640],
+    ["10✅", 640],
+    ["1독도", 640],
 
-    ['1000', 160],
-    ['100✅', 160],
+    ["1000", 160],
+    ["100✅", 160],
 
-    ['10000', 20],
-    ['1000✅', 20],
-    ['fêtes', 20],
-  ])('value %s to be %s', (value, expected) => {
+    ["10000", 20],
+    ["1000✅", 20],
+    ["fêtes", 20],
+  ])("value %s to be %s", (value, expected) => {
     expect(getDomainPriceFromName(value)).toBe(expected);
-  })
+  });
 });
