+++ js/tests/reverse-lookup.test.ts
@@ -1,6 +1,7 @@
 require("dotenv").config();
 import { test, jest, expect } from "@jest/globals";
-import { reverseLookupBatch, reverseLookup } from "../src/utils";
+import { reverseLookupBatch } from "../src/utils/reverseLookupBatch";
+import { reverseLookup } from "../src/utils/reverseLookup";
 import { Connection, PublicKey } from "@solana/web3.js";
 import { performReverseLookupBatch } from "../src/deprecated/utils";
 
@@ -11,10 +12,10 @@ const domain = new PublicKey("Crf8hzfthWGbGbLTVCiqRqV5MVnbpHB1L9KQMd6gsinb");
 
 test("Reverse lookup", async () => {
   performReverseLookupBatch(connection, [domain]).then((e) =>
-    expect(e).toStrictEqual(["bonfida"])
+    expect(e).toStrictEqual(["bonfida"]),
   );
   reverseLookupBatch(connection, [domain]).then((e) =>
-    expect(e).toStrictEqual(["bonfida"])
+    expect(e).toStrictEqual(["bonfida"]),
   );
   reverseLookup(connection, domain).then((e) => expect(e).toBe("bonfida"));
 });
